FIX: moving files from sub to another sub
The problem happens when you have a need to move a sub directory form
one parent to another. ie:

If you have
foo/sub
bar

I want to move foo/sub to bar/sub, problem is bar doesn't exist for some
clients. We have to check for every move if that parent exists.