Add fields to pdf which will help naming
We were finding that the naming of pdf files is important to have as
much data in there as possible so we can use standard command line tools
and file managers to search for specific files. Also to make it easier
to search within pdfcat itself. These extra fields will allow us to
create a naming convention.